pyproject.toml
[tool.poetry]
name = "sailfriend"
version = "0.1.0"
description = ""
authors = ["Raiuri Santos"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.1.1"
djangorestframework = "^3.15.2"
dj-rest-auth = {extras = ["with-social"], version = "^6.0.0"}
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
